requirements.txt
asttokens==2.4.1
branca==0.8.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
earthengine-api==1.2.0
executing==2.1.0
folium==0.18.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
scipy==1.14.1
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
xarray==2024.10.0
xyzservices==2024.9.0
